@@ -34,7 +34,7 @@
     "autoprefixer": "6.5.3",
     "buffer": "5.0.6",
     "check-node-version": "^3.2.0",
-    "copy-webpack-plugin": "4.0.1",
+    "copy-webpack-plugin": "5.0.0",
     "copyfiles": "^2.0.0",
     "core-js": "2.4.1",
     "cross-env": "5.0.5",
@@ -53,28 +53,31 @@
     "file-loader": "0.10.0",
     "fs-extra": "5.0.0",
     "gh-pages": "2.0.1",
+    "html-webpack-plugin": "3.2.0",
     "jasmine": "2.8.0",
     "jasmine-core": "2.8.0",
     "jasmine-node": "2.0.1",
     "jasmine-ts": "0.3.0",
     "jsonfile": "4.0.0",
-    "lerna": "^3.13.0",
+    "lerna": "3.13.0",
     "lodash": "4.17.11",
     "mkdirp": "0.5.1",
     "node-hid": "0.7.3",
     "npm-run-all": "4.0.2",
     "nrf-intel-hex": "1.3.0",
+    "postcss-url": "8.0.0",
     "pre-commit": "1.2.2",
     "request": "2.88.0",
     "rimraf": "2.6.1",
     "source-map-support": "0.5.9",
     "stylelint": "9.6.0",
     "svg-sprite": "1.5.0",
-    "ts-loader": "2.3.1",
+    "ts-loader": "5.3.3",
     "ts-node": "7.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.8.4",
-    "webpack": "3.12.0"
+    "tslint": "5.13.0",
+    "typescript": "3.3.3",
+    "webpack": "4.29.5",
+    "webpack-cli": "3.2.3"
   },
   "pre-commit": [
     "precommit-msg"
