@@ -6,9 +6,9 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "run-p -sn build:*",
-    "build:web": "ng build --prod --aot --base-href=\"\"",
-    "build:renderer": "cross-env AOT_BUILD=true webpack --config webpack.config.js",
-    "server:renderer": "webpack --config webpack.config.js --watch",
+    "build:web": "ng build --prod --aot",
+    "build:renderer": "ng build --prod --aot --project=uhk-renderer",
+    "server:renderer": "ng build --project=uhk-renderer --watch",
     "test-skip": "ng test --watch false",
     "lint": "run-p -snc lint:*",
     "lint:ng": "ng lint",
@@ -19,65 +19,62 @@
   },
   "private": true,
   "devDependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/cli": "1.7.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular-devkit/build-optimizer": "0.3.2",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/language-service": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "@ngrx/effects": "4.0.5",
-    "@ngrx/router-store": "4.0.4",
-    "@ngrx/store": "4.0.3",
-    "@ngrx/store-devtools": "4.0.0",
+    "@angular/animations": "7.2.6",
+    "@angular-builders/custom-webpack": "7.3.1",
+    "@angular/cli": "7.3.2",
+    "@angular/common": "7.2.6",
+    "@angular/compiler": "7.2.6",
+    "@angular/compiler-cli": "7.2.6",
+    "@angular/core": "7.2.6",
+    "@angular-devkit/build-angular": "0.13.3",
+    "@angular-devkit/build-optimizer": "0.13.3",
+    "@angular-devkit/core": "7.3.2",
+    "@angular/forms": "7.2.6",
+    "@angular/http": "7.2.6",
+    "@angular/language-service": "7.2.6",
+    "@angular/platform-browser": "7.2.6",
+    "@angular/platform-browser-dynamic": "7.2.6",
+    "@angular/router": "7.2.6",
+    "@ngtools/webpack": "7.3.3",
+    "@ngrx/effects": "7.2.0",
+    "@ngrx/router-store": "7.2.0",
+    "@ngrx/store": "7.2.0",
+    "@ngrx/store-devtools": "7.2.0",
     "@ngrx/store-log-monitor": "3.0.2",
-    "@types/electron-devtools-installer": "2.0.2",
     "@types/electron-settings": "3.0.0",
     "@types/file-saver": "0.0.1",
     "@types/jasmine": "2.8.8",
     "@types/jasminewd2": "2.0.3",
     "@types/usb": "1.1.3",
-    "angular-confirmation-popover": "3.2.0",
-    "angular-notifier": "2.0.0",
+    "angular-confirmation-popover": "4.2.0",
+    "angular-notifier": "4.1.1",
     "autoprefixer": "^7.2.3",
     "bootstrap": "3.3.7",
-    "buffer": "5.0.6",
-    "circular-dependency-plugin": "^4.2.1",
-    "codelyzer": "3.0.1",
+    "codelyzer": "4.5.0",
     "dragula": "3.7.2",
     "font-awesome": "4.7.0",
-    "html-webpack-plugin": "^2.29.0",
-    "jasmine-core": "3.2.1",
+    "jasmine-core": "3.3.0",
     "jasmine-spec-reporter": "4.2.1",
     "jquery": "3.2.1",
-    "jsonfile": "3.0.1",
-    "karma": "3.0.0",
+    "karma": "3.1.4",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "2.0.1",
+    "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "1.1.2",
     "karma-jasmine-html-reporter": "1.3.1",
-    "ng2-dragula": "1.5.0",
-    "ng2-nouislider": "^1.7.7",
+    "ng2-dragula": "2.1.1",
+    "ng2-nouislider": "1.7.13",
     "ngx-clipboard": "10.0.0",
-    "@ert78gb/ngx-select-ex": "3.7.2",
+    "ngx-select-ex": "3.6.8",
     "ngrx-store-freeze": "0.1.9",
-    "nouislider": "^11.1.0",
-    "postcss-url": "^7.1.2",
+    "nouislider": "13.1.1",
     "protractor": "5.4.0",
-    "rxjs": "5.5.8",
+    "reselect": "3.0.1",
+    "rxjs": "6.4.0",
     "semver": "5.6.0",
     "ts-keycode-enum": "^1.0.6",
     "uhk-common": "1.0.0",
     "xml-loader": "1.2.1",
-    "zone.js": "0.8.26",
-    "@angular-devkit/core": "0.3.2",
-    "@ngtools/webpack": "1.10.2"
+    "zone.js": "0.8.26"
   },
   "dependencies": {
     "classlist.js": "1.1.20150312",
