@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Action, Store } from '@ngrx/store';
-import { Actions, Effect } from '@ngrx/effects';
-import { Observable } from 'rxjs/Observable';
-import { map, startWith, tap, withLatestFrom } from 'rxjs/operators';
+import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Observable } from 'rxjs';
+import { map, mergeMap, startWith, tap, withLatestFrom } from 'rxjs/operators';
 import { NotifierService } from 'angular-notifier';
 
 import { AppStartInfo, LogService, Notification, NotificationType } from 'uhk-common';
@@ -26,8 +26,8 @@ export class ApplicationEffects {
 
     @Effect()
     appStart$: Observable<Action> = this.actions$
-        .ofType(ActionTypes.AppBootstrapped)
         .pipe(
+            ofType(ActionTypes.AppBootstrapped),
             startWith(new AppStartedAction()),
             tap(() => {
                 this.logService.info('Renderer appStart effect start');
@@ -37,17 +37,19 @@ export class ApplicationEffects {
             })
         );
 
-    @Effect({dispatch: false})
+    @Effect({ dispatch: false })
     appStartInfo$: Observable<Action> = this.actions$
-        .ofType(ActionTypes.LoadAppStartInfo)
-        .do(() => {
-            this.appRendererService.getAppStartInfo();
-        });
-
-    @Effect({dispatch: false})
-    showNotification$: Observable<Action> = this.actions$
-        .ofType<ShowNotificationAction>(ActionTypes.AppShowNotification)
         .pipe(
+            ofType(ActionTypes.LoadAppStartInfo),
+            tap(() => {
+                this.appRendererService.getAppStartInfo();
+            })
+        );
+
+    @Effect({ dispatch: false })
+    showNotification$: Observable<Action> = this.actions$
+        .pipe(
+            ofType<ShowNotificationAction>(ActionTypes.AppShowNotification),
             map(action => action.payload),
             tap((notification: Notification) => {
                 if (notification.type === NotificationType.Undoable) {
@@ -59,24 +61,28 @@ export class ApplicationEffects {
 
     @Effect()
     processStartInfo$: Observable<Action> = this.actions$
-        .ofType<ProcessAppStartInfoAction>(ActionTypes.AppProcessStartInfo)
-        .map(action => action.payload)
-        .mergeMap((appInfo: AppStartInfo) => {
-            this.logService.debug('[AppEffect][processStartInfo] payload:', appInfo);
-            return [
-                new ApplyAppStartInfoAction(appInfo),
-                new ConnectionStateChangedAction(appInfo.deviceConnectionState)
-            ];
-        });
+        .pipe(
+            ofType<ProcessAppStartInfoAction>(ActionTypes.AppProcessStartInfo),
+            map(action => action.payload),
+            mergeMap((appInfo: AppStartInfo) => {
+                this.logService.debug('[AppEffect][processStartInfo] payload:', appInfo);
+                return [
+                    new ApplyAppStartInfoAction(appInfo),
+                    new ConnectionStateChangedAction(appInfo.deviceConnectionState)
+                ];
+            })
+        );
 
     @Effect() undoLastNotification$: Observable<Action> = this.actions$
-        .ofType<UndoLastAction>(ActionTypes.UndoLast)
-        .map(action => action.payload)
-        .mergeMap((action: Action) => [action, new DismissUndoNotificationAction()]);
-
-    @Effect({dispatch: false}) openUrlInNewWindow$ = this.actions$
-        .ofType<OpenUrlInNewWindowAction>(ActionTypes.OpenUrlInNewWindow)
         .pipe(
+            ofType<UndoLastAction>(ActionTypes.UndoLast),
+            map(action => action.payload),
+            mergeMap((action: Action) => [action, new DismissUndoNotificationAction()])
+        );
+
+    @Effect({ dispatch: false }) openUrlInNewWindow$ = this.actions$
+        .pipe(
+            ofType<OpenUrlInNewWindowAction>(ActionTypes.OpenUrlInNewWindow),
             withLatestFrom(this.store.select(runningInElectron)),
             tap(([action, inElectron]) => {
                 const url = action.payload;
