@@ -45,7 +45,7 @@ export const initialState: State = {
             firmwareVersion: ''
         }
     },
-    log: [{message: '', cssClass: XtermCssClass.standard}],
+    log: [{ message: '', cssClass: XtermCssClass.standard }],
     restoringUserConfiguration: false,
     hasBackupUserConfiguration: false
 };
@@ -131,7 +131,7 @@ export function reducer(state = initialState, action: Action): State {
                 firmwareUpdateFinished: false,
                 firmwareUpdateFailed: false,
                 firmwareUpdateSuccess: false,
-                log: [{message: 'Start flashing firmware', cssClass: XtermCssClass.standard}]
+                log: [{ message: 'Start flashing firmware', cssClass: XtermCssClass.standard }]
             };
 
         case Device.ActionTypes.UpdateFirmwareSuccess:
@@ -211,7 +211,7 @@ export function reducer(state = initialState, action: Action): State {
             return {
                 ...state,
                 updatingFirmware: true,
-                log: [{message: '', cssClass: XtermCssClass.standard}]
+                log: [{ message: '', cssClass: XtermCssClass.standard }]
             };
         }
         default:
