@@ -7,7 +7,7 @@ export const initialState: Keymap[] = [];
 export function reducer(state = initialState, action: Keymaps.Actions): Keymap[] {
     switch (action.type) {
         case Keymaps.ActionTypes.LoadKeymapsSuccess: {
-            return (action as Keymaps.LoadKeymapSuccessAction).payload ;
+            return (action as Keymaps.LoadKeymapSuccessAction).payload;
         }
 
         default:
